+++ src/Components/Forms/DataGridSelect/DataGridSelect.razor.cs
@@ -254,12 +254,6 @@ public partial class DataGridSelect<TDataItem, TValue> : IDataGrid<TDataItem> wh
          => DataGrid?.OnColumnSortedAsync(column) ?? Task.CompletedTask;
 
     /// <summary>
-    /// Called internally.
-    /// </summary>
-    public Task OnFilterChangedAsync()
-         => DataGrid?.OnFilterChangedAsync() ?? Task.CompletedTask;
-
-    /// <summary>
     /// Removes a column from this grid.
     /// </summary>
     /// <param name="column">The column to remove.</param>